requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic-settings==2.2.1
asyncpg==0.29.0
redis==5.0.1
httpx==0.27.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
cryptography==42.0.2
psycopg2-binary==2.9.9
openai==1.12.0
